+++ frontend/middleware/admin.ts
@@ -1,6 +1,6 @@
 export default defineNuxtRouteMiddleware(async () => {
   /* TODO(grady) add Authorization here
-  const { getMe } = await useMSAL()
+  const { $msal: { getMe } } = useNuxtApp()
   const { isAdmin } = await getMe()
   if (!isAdmin.value) {
     return await navigateTo('/denied')
