+++ frontend/pages/admin/pacta-version/[id].vue
@@ -2,7 +2,7 @@
 import { pactaVersionEditor } from '@/lib/editor'
 
 const router = useRouter()
-const pactaClient = await usePACTA()
+const pactaClient = usePACTA()
 const { loading: { withLoading } } = useModal()
 const { fromParams } = useURLParams()
 const localePath = useLocalePath()
